feat: [CDS-101927]: Categorize CD Code Repositories Connectors
