+++ internal/service/cd_nextgen/connector/codeRepositories/bitbucket.go
@@ -1,4 +1,4 @@
-package connector
+package codeRepositories
 
 import (
 	"context"
@@ -60,14 +60,14 @@ func ResourceConnectorBitbucket() *schema.Resource {
 							AtLeastOneOf:  []string{"api_authentication.0.username", "api_authentication.0.username_ref"},
 						},
 						"username_ref": {
-							Description:   "The name of the Harness secret containing the username." + secret_ref_text,
+							Description:   "The name of the Harness secret containing the username." + secretRefText,
 							Type:          schema.TypeString,
 							Optional:      true,
 							ConflictsWith: []string{"api_authentication.0.username"},
 							AtLeastOneOf:  []string{"api_authentication.0.username", "api_authentication.0.username_ref"},
 						},
 						"token_ref": {
-							Description: "Personal access token for interacting with the BitBucket api." + secret_ref_text,
+							Description: "Personal access token for interacting with the BitBucket api." + secretRefText,
 							Type:        schema.TypeString,
 							Required:    true,
 						},
@@ -98,14 +98,14 @@ func ResourceConnectorBitbucket() *schema.Resource {
 										ExactlyOneOf:  []string{"credentials.0.http.0.username", "credentials.0.http.0.username_ref"},
 									},
 									"username_ref": {
-										Description:   "Reference to a secret containing the username to use for authentication." + secret_ref_text,
+										Description:   "Reference to a secret containing the username to use for authentication." + secretRefText,
 										Type:          schema.TypeString,
 										Optional:      true,
 										ConflictsWith: []string{"credentials.0.http.0.username"},
 										ExactlyOneOf:  []string{"credentials.0.http.0.username", "credentials.0.http.0.username_ref"},
 									},
 									"password_ref": {
-										Description: "Reference to a secret containing the password to use for authentication." + secret_ref_text,
+										Description: "Reference to a secret containing the password to use for authentication." + secretRefText,
 										Type:        schema.TypeString,
 										Optional:    true,
 									},
@@ -122,7 +122,7 @@ func ResourceConnectorBitbucket() *schema.Resource {
 							Elem: &schema.Resource{
 								Schema: map[string]*schema.Schema{
 									"ssh_key_ref": {
-										Description: "Reference to the Harness secret containing the ssh key." + secret_ref_text,
+										Description: "Reference to the Harness secret containing the ssh key." + secretRefText,
 										Type:        schema.TypeString,
 										Required:    true,
 									},
