+++ internal/service/cd_nextgen/connector/codeRepositories/bitbucket_data_source.go
@@ -1,4 +1,4 @@
-package connector
+package codeRepositories
 
 import (
 	"fmt"
@@ -9,19 +9,19 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 )
 
-func DatasourceConnectorGitlab() *schema.Resource {
+func DatasourceConnectorBitbucket() *schema.Resource {
 	resource := &schema.Resource{
-		Description: "Datasource for looking up a Gitlab connector.",
-		ReadContext: resourceConnectorGitlabRead,
+		Description: "Datasource for looking up a Bitbucket connector.",
+		ReadContext: resourceConnectorBitbucketRead,
 
 		Schema: map[string]*schema.Schema{
 			"url": {
-				Description: "URL of the gitlab repository or account.",
+				Description: "URL of the BitBucket repository or account.",
 				Type:        schema.TypeString,
 				Computed:    true,
 			},
 			"connection_type": {
-				Description: fmt.Sprintf("Whether the connection we're making is to a gitlab repository or a gitlab account. Valid values are %s.", strings.Join(nextgen.GitConnectorTypeValues, ", ")),
+				Description: fmt.Sprintf("Whether the connection we're making is to a BitBucket repository or a BitBucket account. Valid values are %s.", strings.Join(nextgen.GitConnectorTypeValues, ", ")),
 				Type:        schema.TypeString,
 				Computed:    true,
 			},
@@ -37,13 +37,23 @@ func DatasourceConnectorGitlab() *schema.Resource {
 				Elem:        &schema.Schema{Type: schema.TypeString},
 			},
 			"api_authentication": {
-				Description: "Configuration for using the gitlab api. API Access is Computed for using “Git Experience”, for creation of Git based triggers, Webhooks management and updating Git statuses.",
+				Description: "Configuration for using the BitBucket api. API Access is required for using “Git Experience”, for creation of Git based triggers, Webhooks management and updating Git statuses.",
 				Type:        schema.TypeList,
 				Computed:    true,
 				Elem: &schema.Resource{
 					Schema: map[string]*schema.Schema{
+						"username": {
+							Description: "The username used for connecting to the api.",
+							Type:        schema.TypeString,
+							Computed:    true,
+						},
+						"username_ref": {
+							Description: "The name of the Harness secret containing the username." + secretRefText,
+							Type:        schema.TypeString,
+							Computed:    true,
+						},
 						"token_ref": {
-							Description: "Personal access token for interacting with the gitlab api." + secret_ref_text,
+							Description: "Personal access token for interacting with the BitBucket api." + secretRefText,
 							Type:        schema.TypeString,
 							Computed:    true,
 						},
@@ -68,17 +78,12 @@ func DatasourceConnectorGitlab() *schema.Resource {
 										Computed:    true,
 									},
 									"username_ref": {
-										Description: "Reference to a secret containing the username to use for authentication." + secret_ref_text,
-										Type:        schema.TypeString,
-										Computed:    true,
-									},
-									"token_ref": {
-										Description: "Reference to a secret containing the personal access to use for authentication." + secret_ref_text,
+										Description: "Reference to a secret containing the username to use for authentication." + secretRefText,
 										Type:        schema.TypeString,
 										Computed:    true,
 									},
 									"password_ref": {
-										Description: "Reference to a secret containing the password to use for authentication." + secret_ref_text,
+										Description: "Reference to a secret containing the password to use for authentication." + secretRefText,
 										Type:        schema.TypeString,
 										Computed:    true,
 									},
@@ -92,7 +97,7 @@ func DatasourceConnectorGitlab() *schema.Resource {
 							Elem: &schema.Resource{
 								Schema: map[string]*schema.Schema{
 									"ssh_key_ref": {
-										Description: "Reference to the Harness secret containing the ssh key." + secret_ref_text,
+										Description: "Reference to the Harness secret containing the ssh key." + secretRefText,
 										Type:        schema.TypeString,
 										Computed:    true,
 									},
