+++ internal/service/cd_nextgen/connector/codeRepositories/bitbucket_test.go
@@ -1,4 +1,4 @@
-package connector_test
+package codeRepositories_test
 
 import (
 	"fmt"
@@ -9,12 +9,12 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
 )
 
-func TestAccResourceConnectorGit_Http(t *testing.T) {
+func TestAccResourceConnectorBitbucket_Http(t *testing.T) {
 
 	id := fmt.Sprintf("%s_%s", t.Name(), utils.RandStringBytes(5))
 	name := id
 	updatedName := fmt.Sprintf("%s_updated", name)
-	resourceName := "harness_platform_connector_git.test"
+	resourceName := "harness_platform_connector_bitbucket.test"
 
 	resource.UnitTest(t, resource.TestCase{
 		PreCheck:          func() { acctest.TestAccPreCheck(t) },
@@ -25,14 +25,14 @@ func TestAccResourceConnectorGit_Http(t *testing.T) {
 		CheckDestroy: testAccConnectorDestroy(resourceName),
 		Steps: []resource.TestStep{
 			{
-				Config: testAccResourceConnectorGit_http(id, name),
+				Config: testAccResourceConnectorBitbucket_http(id, name),
 				Check: resource.ComposeTestCheckFunc(
 					resource.TestCheckResourceAttr(resourceName, "id", id),
 					resource.TestCheckResourceAttr(resourceName, "identifier", id),
 					resource.TestCheckResourceAttr(resourceName, "name", name),
 					resource.TestCheckResourceAttr(resourceName, "description", "test"),
 					resource.TestCheckResourceAttr(resourceName, "tags.#", "1"),
-					resource.TestCheckResourceAttr(resourceName, "url", "https://git.example.com/account"),
+					resource.TestCheckResourceAttr(resourceName, "url", "https://bitbucket.com/account"),
 					resource.TestCheckResourceAttr(resourceName, "connection_type", "Account"),
 					resource.TestCheckResourceAttr(resourceName, "validation_repo", "some_repo"),
 					resource.TestCheckResourceAttr(resourceName, "delegate_selectors.#", "1"),
@@ -40,7 +40,7 @@ func TestAccResourceConnectorGit_Http(t *testing.T) {
 				),
 			},
 			{
-				Config: testAccResourceConnectorGit_http(id, updatedName),
+				Config: testAccResourceConnectorBitbucket_http(id, updatedName),
 				Check: resource.ComposeTestCheckFunc(
 					resource.TestCheckResourceAttr(resourceName, "id", id),
 					resource.TestCheckResourceAttr(resourceName, "identifier", id),
@@ -59,29 +59,29 @@ func TestAccResourceConnectorGit_Http(t *testing.T) {
 	})
 }
 
-func TestAccResourceConnectorGit_Ssh(t *testing.T) {
+func TestAccResourceConnectorBitbucket_Ssh(t *testing.T) {
 
 	id := fmt.Sprintf("%s_%s", t.Name(), utils.RandStringBytes(5))
 	name := id
-	resourceName := "harness_platform_connector_git.test"
+	resourceName := "harness_platform_connector_bitbucket.test"
 
 	resource.UnitTest(t, resource.TestCase{
 		PreCheck:          func() { acctest.TestAccPreCheck(t) },
 		ProviderFactories: acctest.ProviderFactories,
+		CheckDestroy:      testAccConnectorDestroy(resourceName),
 		ExternalProviders: map[string]resource.ExternalProvider{
 			"time": {},
 		},
-		CheckDestroy: testAccConnectorDestroy(resourceName),
 		Steps: []resource.TestStep{
 			{
-				Config: testAccResourceConnectorGit_ssh(id, name),
+				Config: testAccResourceConnectorBitbucket_ssh(id, name),
 				Check: resource.ComposeTestCheckFunc(
 					resource.TestCheckResourceAttr(resourceName, "id", id),
 					resource.TestCheckResourceAttr(resourceName, "identifier", id),
 					resource.TestCheckResourceAttr(resourceName, "name", name),
 					resource.TestCheckResourceAttr(resourceName, "description", "test"),
 					resource.TestCheckResourceAttr(resourceName, "tags.#", "1"),
-					resource.TestCheckResourceAttr(resourceName, "url", "https://git.example.com/account"),
+					resource.TestCheckResourceAttr(resourceName, "url", "https://bitbucket.com/account"),
 					resource.TestCheckResourceAttr(resourceName, "connection_type", "Account"),
 					resource.TestCheckResourceAttr(resourceName, "validation_repo", "some_repo"),
 					resource.TestCheckResourceAttr(resourceName, "delegate_selectors.#", "1"),
@@ -96,7 +96,59 @@ func TestAccResourceConnectorGit_Ssh(t *testing.T) {
 	})
 }
 
-func testAccResourceConnectorGit_http(id string, name string) string {
+func TestAccResourceConnectorBitbucket_api_token(t *testing.T) {
+
+	id := fmt.Sprintf("%s_%s", t.Name(), utils.RandStringBytes(5))
+	name := id
+	updatedName := fmt.Sprintf("%s_updated", name)
+	resourceName := "harness_platform_connector_bitbucket.test"
+
+	resource.UnitTest(t, resource.TestCase{
+		PreCheck:          func() { acctest.TestAccPreCheck(t) },
+		ProviderFactories: acctest.ProviderFactories,
+		ExternalProviders: map[string]resource.ExternalProvider{
+			"time": {},
+		},
+		CheckDestroy: testAccConnectorDestroy(resourceName),
+		Steps: []resource.TestStep{
+			{
+				Config: testAccResourceConnectorBitbucket_api_token(id, name),
+				Check: resource.ComposeTestCheckFunc(
+					resource.TestCheckResourceAttr(resourceName, "id", id),
+					resource.TestCheckResourceAttr(resourceName, "identifier", id),
+					resource.TestCheckResourceAttr(resourceName, "name", name),
+					resource.TestCheckResourceAttr(resourceName, "description", "test"),
+					resource.TestCheckResourceAttr(resourceName, "tags.#", "1"),
+					resource.TestCheckResourceAttr(resourceName, "url", "https://bitbucket.com/account"),
+					resource.TestCheckResourceAttr(resourceName, "connection_type", "Account"),
+					resource.TestCheckResourceAttr(resourceName, "validation_repo", "some_repo"),
+					resource.TestCheckResourceAttr(resourceName, "delegate_selectors.#", "1"),
+					resource.TestCheckResourceAttr(resourceName, "credentials.0.http.0.username", "admin"),
+					resource.TestCheckResourceAttr(resourceName, "api_authentication.0.username", "admin"),
+				),
+			},
+			{
+				Config: testAccResourceConnectorBitbucket_api_token(id, updatedName),
+				Check: resource.ComposeTestCheckFunc(
+					resource.TestCheckResourceAttr(resourceName, "id", id),
+					resource.TestCheckResourceAttr(resourceName, "identifier", id),
+					resource.TestCheckResourceAttr(resourceName, "name", updatedName),
+					resource.TestCheckResourceAttr(resourceName, "description", "test"),
+					resource.TestCheckResourceAttr(resourceName, "tags.#", "1"),
+					resource.TestCheckResourceAttr(resourceName, "credentials.0.http.0.username", "admin"),
+					resource.TestCheckResourceAttr(resourceName, "api_authentication.0.username", "admin"),
+				),
+			},
+			{
+				ResourceName:      resourceName,
+				ImportState:       true,
+				ImportStateVerify: true,
+			},
+		},
+	})
+}
+
+func testAccResourceConnectorBitbucket_http(id string, name string) string {
 	return fmt.Sprintf(`
 	resource "harness_platform_secret_text" "test" {
 		identifier = "%[1]s"
@@ -109,13 +161,13 @@ func testAccResourceConnectorGit_http(id string, name string) string {
 		value = "secret"
 	}
 
-		resource "harness_platform_connector_git" "test" {
+		resource "harness_platform_connector_bitbucket" "test" {
 			identifier = "%[1]s"
 			name = "%[2]s"
 			description = "test"
 			tags = ["foo:bar"]
 
-			url = "https://git.example.com/account"
+			url = "https://bitbucket.com/account"
 			connection_type = "Account"
 			validation_repo = "some_repo"
 			delegate_selectors = ["harness-delegate"]
@@ -135,7 +187,7 @@ func testAccResourceConnectorGit_http(id string, name string) string {
 `, id, name)
 }
 
-func testAccResourceConnectorGit_ssh(id string, name string) string {
+func testAccResourceConnectorBitbucket_api_token(id string, name string) string {
 	return fmt.Sprintf(`
 	resource "harness_platform_secret_text" "test" {
 		identifier = "%[1]s"
@@ -148,13 +200,57 @@ func testAccResourceConnectorGit_ssh(id string, name string) string {
 		value = "secret"
 	}
 
-		resource "harness_platform_connector_git" "test" {
+		resource "harness_platform_connector_bitbucket" "test" {
 			identifier = "%[1]s"
 			name = "%[2]s"
 			description = "test"
 			tags = ["foo:bar"]
 
-			url = "https://git.example.com/account"
+			url = "https://bitbucket.com/account"
+			connection_type = "Account"
+			validation_repo = "some_repo"
+			delegate_selectors = ["harness-delegate"]
+			credentials {
+				http {
+					username = "admin"
+					password_ref = "account.${harness_platform_secret_text.test.id}"
+				}
+			}
+
+			api_authentication {
+				username = "admin"
+				token_ref = "account.${harness_platform_secret_text.test.id}"
+			}
+			depends_on = [time_sleep.wait_4_seconds]
+		}
+
+		resource "time_sleep" "wait_4_seconds" {
+			depends_on = [harness_platform_secret_text.test]
+			destroy_duration = "4s"
+		}
+`, id, name)
+}
+
+func testAccResourceConnectorBitbucket_ssh(id string, name string) string {
+	return fmt.Sprintf(`
+	resource "harness_platform_secret_text" "test" {
+		identifier = "%[1]s"
+		name = "%[2]s"
+		description = "test"
+		tags = ["foo:bar"]
+
+		secret_manager_identifier = "harnessSecretManager"
+		value_type = "Inline"
+		value = "secret"
+	}
+
+		resource "harness_platform_connector_bitbucket" "test" {
+			identifier = "%[1]s"
+			name = "%[2]s"
+			description = "test"
+			tags = ["foo:bar"]
+
+			url = "https://bitbucket.com/account"
 			connection_type = "Account"
 			validation_repo = "some_repo"
 			delegate_selectors = ["harness-delegate"]
@@ -163,6 +259,7 @@ func testAccResourceConnectorGit_ssh(id string, name string) string {
 					ssh_key_ref = "account.${harness_platform_secret_text.test.id}"
 				}
 			}
+
 			depends_on = [time_sleep.wait_4_seconds]
 		}
