+++ internal/service/cd_nextgen/connector/codeRepositories/github_data_source_test.go
@@ -1,4 +1,4 @@
-package connector_test
+package codeRepositories_test
 
 import (
 	"fmt"
