+++ internal/service/cd_nextgen/connector/codeRepositories/gitlab_data_source.go
@@ -1,4 +1,4 @@
-package connector
+package codeRepositories
 
 import (
 	"fmt"
@@ -9,19 +9,19 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 )
 
-func DatasourceConnectorGit() *schema.Resource {
+func DatasourceConnectorGitlab() *schema.Resource {
 	resource := &schema.Resource{
-		Description: "Datasource for looking up a Git connector.",
-		ReadContext: resourceConnectorGitRead,
+		Description: "Datasource for looking up a Gitlab connector.",
+		ReadContext: resourceConnectorGitlabRead,
 
 		Schema: map[string]*schema.Schema{
 			"url": {
-				Description: "URL of the git repository or account.",
+				Description: "URL of the gitlab repository or account.",
 				Type:        schema.TypeString,
 				Computed:    true,
 			},
 			"connection_type": {
-				Description: fmt.Sprintf("Whether the connection we're making is to a git repository or a git account. Valid values are %s.", strings.Join(nextgen.GitConnectorTypeValues, ", ")),
+				Description: fmt.Sprintf("Whether the connection we're making is to a gitlab repository or a gitlab account. Valid values are %s.", strings.Join(nextgen.GitConnectorTypeValues, ", ")),
 				Type:        schema.TypeString,
 				Computed:    true,
 			},
@@ -35,6 +35,20 @@ func DatasourceConnectorGit() *schema.Resource {
 				Type:        schema.TypeSet,
 				Computed:    true,
 				Elem:        &schema.Schema{Type: schema.TypeString},
+			},
+			"api_authentication": {
+				Description: "Configuration for using the gitlab api. API Access is Computed for using “Git Experience”, for creation of Git based triggers, Webhooks management and updating Git statuses.",
+				Type:        schema.TypeList,
+				Computed:    true,
+				Elem: &schema.Resource{
+					Schema: map[string]*schema.Schema{
+						"token_ref": {
+							Description: "Personal access token for interacting with the gitlab api." + secretRefText,
+							Type:        schema.TypeString,
+							Computed:    true,
+						},
+					},
+				},
 			},
 			"credentials": {
 				Description: "Credentials to use for the connection.",
@@ -54,12 +68,17 @@ func DatasourceConnectorGit() *schema.Resource {
 										Computed:    true,
 									},
 									"username_ref": {
-										Description: "Reference to a secret containing the username to use for authentication." + secret_ref_text,
+										Description: "Reference to a secret containing the username to use for authentication." + secretRefText,
+										Type:        schema.TypeString,
+										Computed:    true,
+									},
+									"token_ref": {
+										Description: "Reference to a secret containing the personal access to use for authentication." + secretRefText,
 										Type:        schema.TypeString,
 										Computed:    true,
 									},
 									"password_ref": {
-										Description: "Reference to a secret containing the password to use for authentication." + secret_ref_text,
+										Description: "Reference to a secret containing the password to use for authentication." + secretRefText,
 										Type:        schema.TypeString,
 										Computed:    true,
 									},
@@ -73,7 +92,7 @@ func DatasourceConnectorGit() *schema.Resource {
 							Elem: &schema.Resource{
 								Schema: map[string]*schema.Schema{
 									"ssh_key_ref": {
-										Description: "Reference to the Harness secret containing the ssh key." + secret_ref_text,
+										Description: "Reference to the Harness secret containing the ssh key." + secretRefText,
 										Type:        schema.TypeString,
 										Computed:    true,
 									},
