+++ internal/service/cd_nextgen/connector/codeRepositories/secret_util.go
@@ -1,0 +1,3 @@
+package codeRepositories
+
+const secretRefText = " To reference a secret at the organization scope, prefix 'org' to the expression: org.{identifier}. To reference a secret at the account scope, prefix 'account` to the expression: account.{identifier}."
